+++ spec/extra_core/symbol_spec.rb
@@ -1,0 +1,15 @@
+if RUBY_PLATFORM == 'opal'
+else
+  require 'spec_helper'
+  require 'volt/extra_core/symbol'
+
+  describe Symbol do
+    it 'should pluralize correctly' do
+      expect(:car.pluralize).to eq(:cars)
+    end
+
+    it 'should singularize correctly' do
+      expect(:cars.singularize).to eq(:car)
+    end
+  end
+end